+++ packages/jest-jasmine2/src/index.ts
@@ -180,7 +180,7 @@ export default async function jasmine2(
     if (esm) {
       await runtime.unstable_importModule(path);
     } else {
-      const setupFile = runtime.requireModule(path);
+      const setupFile = await runtime.requireModuleWithEsmPreload(path);
       if (typeof setupFile === 'function') {
         await setupFile();
       }
@@ -196,7 +196,7 @@ export default async function jasmine2(
   if (esm) {
     await runtime.unstable_importModule(testPath);
   } else {
-    runtime.requireModule(testPath);
+    await runtime.requireModuleWithEsmPreload(testPath);
   }
 
   await env.execute();